add functionality for the "complete" button for todos list, add login functionality

+++ TodoApp/routers/todos.py
@@ -14,6 +14,8 @@ from database import engine, SessionLocal
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 
+from .auth import get_current_user
+
 router = APIRouter(
     prefix="/todos",
     tags=["todos"],
@@ -34,25 +36,37 @@ def get_db():
 
 @router.get("/", response_class=HTMLResponse)
 async def read_all_by_user(request: Request, db: Session = Depends(get_db)):
-    todos = db.query(models.Todos).filter(models.Todos.owner_id == 1).all()
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
+    todos = db.query(models.Todos).filter(models.Todos.owner_id == user.get("id")).all()
 
     return templates.TemplateResponse("home.html", {"request": request, "todos": todos})
 
 
 @router.get("/add-todo", response_class=HTMLResponse)
 async def add_new_todo(request: Request):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     return templates.TemplateResponse("add-todo.html", {"request": request})
 
 
 @router.post("/add-todo", response_class=HTMLResponse)
 async def create_todo(request: Request, title: str = Form(...), description: str = Form(...),
                       priority: int = Form(...), db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_model = models.Todos()
     todo_model.title = title
     todo_model.description = description
     todo_model.priority = priority
     todo_model.complete = False
-    todo_model.owner_id = 1
+    todo_model.owner_id = user.get('id')
 
     db.add(todo_model)
     db.commit()
@@ -62,6 +76,10 @@ async def create_todo(request: Request, title: str = Form(...), description: str
 
 @router.get("/edit-todo/{todo_id}", response_class=HTMLResponse)
 async def edit_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
 
     return templates.TemplateResponse("edit-todo.html", {"request": request, "todo": todo})
@@ -70,6 +88,10 @@ async def edit_todo(request: Request, todo_id: int, db: Session = Depends(get_db
 @router.post("/edit-todo/{todo_id}", response_class=HTMLResponse)
 async def edit_todo_commit(request: Request, todo_id: int, title: str = Form(...), description: str = Form(...),
                            priority: int = Form(...), db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
     todo_model.title = title
     todo_model.description = description
@@ -83,8 +105,12 @@ async def edit_todo_commit(request: Request, todo_id: int, title: str = Form(...
 
 @router.get("/delete/{todo_id}")
 async def delete_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id)\
-        .filter(models.Todos.owner_id == 1).first()
+        .filter(models.Todos.owner_id == user.get('id')).first()
 
     if todo_model:
         db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
@@ -92,3 +118,16 @@ async def delete_todo(request: Request, todo_id: int, db: Session = Depends(get_
     return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
 
 
+@router.get("/complete/{todo_id}", response_class=HTMLResponse)
+async def complete_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
+    todo = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+    todo.complete = not todo.complete
+
+    db.add(todo)
+    db.commit()
+
+    return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
